pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "knowledgenet"
version = "1.0.0b1"
authors = [
  { name="Amit Chatterjee"},
]

dependencies = [
  "opentelemetry-api >= 1.39.1",
  "opentelemetry-sdk >= 1.39.1",
  "opentelemetry-exporter-otlp >= 1.39.1"
]

description = "Knowledgenet RETE engine"
requires-python = ">=3.13"
classifiers = [
    "Programming Language :: Python :: 3",
    "Operating System :: OS Independent",
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
]
keywords = [
    "RETE",
    "rules engine",
    "knowledgebase",
    "inference engine",
    "expert systems",
    "rule-based system",
    "forward chaining",
    "backward chaining",
]
license = "Apache-2.0"
readme = "readme.md"
[project.urls]
Homepage = "https://github.com/amitchatterjee/knowledgenet"
Issues = "https://github.com/amitchatterjee/knowledgenet"
Repository = "https://github.com/amitchatterjee/knowledgenet.git"
Documentation = "https://github.com/amitchatterjee/knowledgenet/tree/develop/doc"
